% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcGseaStatCumulative}
\alias{calcGseaStatCumulative}
\title{Calculates GSEA statistic values for all the prefixes of a gene set}
\usage{
calcGseaStatCumulative(stats, selectedStats, gseaParam)
}
\arguments{
\item{stats}{Named numeric vector with gene-level statistics}

\item{selectedStats}{indexes of selected genes in a 'stats' array}

\item{gseaParam}{GSEA weight parameter (0 is unweighted, suggested value is 1)}
}
\value{
numeric vector of GSEA statistics for all prefixes of selectedStats
}
\description{
Calculates GSEA statistic values for all the prefixes of a gene set
}

