% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geseca-multilevel.R
\name{collapsePathwaysGeseca}
\alias{collapsePathwaysGeseca}
\title{Collapse list of enriched pathways to independent ones (GESECA version).}
\usage{
collapsePathwaysGeseca(
  gesecaRes,
  pathways,
  E,
  eps = 1e-50,
  checkDepth = 10,
  nproc = 0,
  BPPARAM = NULL
)
}
\arguments{
\item{gesecaRes}{Table with results of running geseca(), should be filtered
by p-value, for example by selecting ones with padj < 0.01.}

\item{pathways}{List of pathways, should contain all the pathways present in
`gesecaRes`.}

\item{E}{expression matrix, the same as in `geseca()`.}

\item{center}{a logical value indicating whether the gene expression should be centered to have zero mean before the analysis takes place.
The default is TRUE. The value is passed to \link[base]{scale}.}

\item{scale}{a logical value indicating whether the gene expression should be scaled to have
unit variance before the analysis takes place.
The default is FALSE. The value is passed to \link[base]{scale}.}

\item{pval.threshold}{Two pathways are considered dependent when p-value
of enrichment of one pathways on background of another
is greater then `pval.threshold`.}

\item{nperm}{Number of permutations to test for independence, should be
several times greater than `1/pval.threhold`.
Default value: `10/pval.threshold`.}
}
\value{
Named list with two elments: `mainPathways` containing IDs of pathways
        not reducable to each other, and `parentPathways` with vector describing
        for all the pathways to which ones they can be reduced. For
        pathways from `mainPathwyas` vector `parentPathways` contains `NA` values.
}
\description{
Collapse list of enriched pathways to independent ones (GESECA version).
}
