% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea-package.R
\docType{data}
\name{examplePathways}
\alias{examplePathways}
\title{Example list of mouse Reactome pathways.}
\description{
The list was obtained by selecting all the pathways from `reactome.db`
package that contain mouse genes. The exact script is available as
system.file("gen_reactome_pathways.R", package="fgsea")
}
