% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea-package.R
\docType{data}
\name{exampleRanks}
\alias{exampleRanks}
\title{Example vector of gene-lvele statistics obtained for Th1 polarization}
\description{
Example vector of gene-lvele statistics obtained for Th1 polarization
}

