% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea.R
\name{fgsea}
\alias{fgsea}
\title{Runs preranked gene set enrichment analysis.}
\usage{
fgsea(...)
}
\arguments{
\item{...}{optional arguments for functions \link[fgsea]{fgseaSimple}, \link[fgsea]{fgseaMultilevel}}
}
\value{
A table with GSEA results. Each row corresponds to a tested pathway.
}
\description{
Runs preranked gene set enrichment analysis.
}
\examples{
data(examplePathways)
data(exampleRanks)
fgseaRes <- fgsea(examplePathways, exampleRanks, maxSize=500)
# Testing only one pathway is implemented in a more efficient manner
fgseaRes1 <- fgsea(examplePathways[1], exampleRanks)
}
