% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgseaMultilevel.R
\name{fgseaMultilevel}
\alias{fgseaMultilevel}
\title{Runs preranked gene set enrichment analysis.}
\usage{
fgseaMultilevel(pathways, stats, samplesSize = 101, minSize = 1,
  maxSize = Inf, absEps = 0, signed = FALSE, nproc = 0,
  BPPARAM = NULL)
}
\arguments{
\item{pathways}{List of gene sets to check.}

\item{stats}{Named vector of gene-level stats. Names should be the same as in 'pathways'}

\item{samplesSize}{The size of a random set of genes which in turn has size = pathwaySize}

\item{minSize}{Minimal size of a gene set to test. All pathways below the threshold are excluded.}

\item{maxSize}{Maximal size of a gene set to test. All pathways above the threshold are excluded.}

\item{absEps}{This parameter sets the boundary for calculating the p value.}

\item{signed}{If TRUE returns two probabilities with a sign dependency.}

\item{nproc}{If not equal to zero sets BPPARAM to use nproc workers (default = 0).}

\item{BPPARAM}{Parallelization parameter used in bplapply.}
}
\value{
A table with GSEA results. Each row corresponds to a tested pathway.
}
\description{
This approach is based on multilevel splitting approach,
that server for estimating the probability of rare events.
}
\examples{
data(examplePathways)
data(exampleRanks)
fgseaMultilevelRes <- fgseaMultilevel(examplePathways, exampleRanks, maxSize=500)
}
