% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fgseaMultilevelCpp}
\alias{fgseaMultilevelCpp}
\title{Calculates low GSEA p-values for a given gene set size using the multilevel split Monte Carlo approach.}
\usage{
fgseaMultilevelCpp(
  enrichmentScores,
  ranks,
  pathwaySize,
  sampleSize,
  seed,
  eps,
  sign,
  moveScale = 1,
  logStatus = FALSE
)
}
\arguments{
\item{enrichmentScores}{A vector of enrichment scores, for which p-values should be calculated}

\item{ranks}{An integer vector with the gene-level statistics}

\item{pathwaySize}{A scalar with the size of the gene set}

\item{seed}{Random seed}

\item{eps}{P-values below eps aren't calculated}

\item{sign}{Controls whether ES^+ or ES score is used}

\item{moveScale}{Controls the number of MCMC iterations on each level}

\item{logStatus}{Controls whether debug output should be shown}
}
\value{
table with p-values and estimation errors
}
\description{
Calculates low GSEA p-values for a given gene set size using the multilevel split Monte Carlo approach.
}
\keyword{internal}
