% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgseaMultilevel.R
\name{multilevelError}
\alias{multilevelError}
\title{Calculates the expected error.}
\usage{
multilevelError(pval, C)
}
\arguments{
\item{pval}{P-value}

\item{C}{equivavlent to samplesSize}
}
\value{
The value of the expected error
}
\description{
Calculates the expected error.
}
\examples{
expectedError <- multilevelError(pval=1e-10, C=1001)
}
