% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgseaMultilevel.R
\name{multilevelImpl}
\alias{multilevelImpl}
\title{Calculates P-values for preprocessed data.}
\usage{
multilevelImpl(
  multilevelPathwaysList,
  stats,
  sampleSize,
  seed,
  eps,
  sign = FALSE,
  BPPARAM = NULL
)
}
\arguments{
\item{multilevelPathwaysList}{List of pathways for which P-values will be calculated.}

\item{stats}{Named vector of gene-level stats. Names should be the same as in 'pathways'}

\item{sampleSize}{The size of a random set of genes which in turn has size = pathwaySize}

\item{seed}{`seed` parameter from `fgseaMultilevel`}

\item{eps}{This parameter sets the boundary for calculating the p value.}

\item{sign}{This option will be used in future implementations.}

\item{BPPARAM}{Parallelization parameter used in bplapply.
Can be used to specify cluster to run. If not initialized explicitly or
by setting `nproc` default value `bpparam()` is used.}
}
\value{
List of P-values.
}
\description{
Calculates P-values for preprocessed data.
}
