% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geseca-plot.R
\name{plotCoregulationProfile}
\alias{plotCoregulationProfile}
\title{Plots expression profile of a gene set}
\usage{
plotCoregulationProfile(pathway, E, titles = colnames(E), conditions = NULL)
}
\arguments{
\item{pathway}{Gene set to plot.}

\item{E}{matrix with gene expression values}

\item{titles}{sample titles to use for labels}

\item{conditions}{sample grouping to use for coloring}
}
\value{
ggplot object with the coregulation profile plot
}
\description{
Plots expression profile of a gene set
}
