% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geseca-plot.R
\name{plotCoregulationProfileReduction}
\alias{plotCoregulationProfileReduction}
\title{Plot a spatial expression profile of a gene set}
\usage{
plotCoregulationProfileReduction(
  pathway,
  object,
  title = NULL,
  assay = DefaultAssay(object),
  reduction = NULL,
  colors = c("darkblue", "lightgrey", "darkred"),
  guide = "colourbar",
  ...
)
}
\arguments{
\item{pathway}{Gene set to plot or a list of gene sets (see details below)}

\item{object}{Seurat object}

\item{title}{plot title}

\item{assay}{assay to use for obtaining scaled data, preferably with}

\item{reduction}{reduction to use for plotting (one of the `Seurat::Reductions(object)`)}

\item{colors}{vector of three colors to use in the color scheme}

\item{guide}{option for `ggplot2::scale_color_gradientn` to control for presence of the color legend
the same universe of genes in the scaled data}

\item{...}{additional arguments for Seurat::FeaturePlot}
}
\value{
ggplot object (or a list of objects) with the coregulation profile plot

When the input is a list of pathways, pathway names are used for titles.
A list of ggplot objects a returned in that case.
}
\description{
Plot a spatial expression profile of a gene set
}
