% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geseca-plot.R
\name{plotGesecaTable}
\alias{plotGesecaTable}
\title{Plots table of gene set profiles.}
\usage{
plotGesecaTable(
  pathways,
  E,
  gesecaRes,
  colwidths = c(5, 3, 0.8, 1.2, 1.2),
  titles = colnames(E),
  pathwayLabelStyle = NULL,
  headerLabelStyle = NULL,
  valueStyle = NULL,
  axisLabelStyle = NULL,
  axisLabelHeightScale = NULL
)
}
\arguments{
\item{pathways}{Pathways to plot table, as in `geseca` function.}

\item{E}{gene expression matrix, as in `geseca` function.}

\item{gesecaRes}{Table with geseca results.}

\item{colwidths}{Vector of five elements corresponding to column width for
grid.arrange. Can be both units and simple numeric vector, in latter case
it defines proportions, not actual sizes. If column width is set to zero, the column is not drawn.}

\item{pathwayLabelStyle}{list with style parameter adjustments for pathway labels.
For example, `list(size=10, color="red")` set the font size to 10 and color to red.
See `cowplot::draw_text` for possible options.}

\item{headerLabelStyle}{similar to `pathwayLabelStyle` but for the table header.}

\item{valueStyle}{similar to `pathwayLabelStyle` but for pctVar and p-value columns.}

\item{axisLabelStlye}{list with style parameter adjustments for sample labels.
See `ggplot2::element_text` for possible options.}
}
\value{
ggplot object with gene set profile plots
}
\description{
Plots table of gene set profiles.
}
