% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.R
\name{reactomePathways}
\alias{reactomePathways}
\title{Returns a list of Reactome pathways for given Entrez gene IDs}
\usage{
reactomePathways(genes)
}
\arguments{
\item{genes}{Entrez IDs of query genes.}
}
\value{
A list of vectors with gene sets.
}
\description{
Returns a list of Reactome pathways for given Entrez gene IDs
}
\examples{
data(exampleRanks)
pathways <- reactomePathways(names(exampleRanks))
}
