% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.R
\name{writeGmtPathways}
\alias{writeGmtPathways}
\title{Write collection of pathways (list of vectors) to a gmt file}
\usage{
writeGmtPathways(pathways, gmt.file)
}
\arguments{
\item{pathways}{a named list of vectors with gene ids}

\item{gmt.file}{name of the output file}
}
\description{
Write collection of pathways (list of vectors) to a gmt file
}
\examples{
data(examplePathways)
writeGmtPathways(examplePathways, tempfile("examplePathways", fileext=".gmt"))
}
