% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea.R
\name{calcGseaStat}
\alias{calcGseaStat}
\title{Calculates GSEA statistics for a given query gene set}
\usage{
calcGseaStat(stats, selectedStats, gseaParam = 1)
}
\arguments{
\item{stats}{Named numeric vector with gene-level statistics}

\item{selectedStats}{indexes of selected genes in a 'stats' array}

\item{gseaParam}{GSEA weight parameter (0 is unweighted, suggested value is 1)}
}
\value{
value of GSEA statistic
}
\description{
Calculates GSEA statistics for a given query gene set
}

