% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcGseaStatCumulative}
\alias{calcGseaStatCumulative}
\title{Calculates GSEA statistic values for all the prefixes of a gene set}
\usage{
calcGseaStatCumulative(stats, selectedStats, gseaParam)
}
\arguments{
\item{stats}{Named numeric vector with gene-level statistics
sorted in decreasing order (order is not checked)}

\item{selectedStats}{indexes of selected genes in a `stats` array}

\item{gseaParam}{GSEA weight parameter (0 is unweighted, suggested value is 1)}
}
\value{
Numeric vector of GSEA statistics for all prefixes of selectedStats.
}
\description{
Takes \emph{O(k^\{3/2\})} time, where \emph{k} is a size of `selectedSize`.
}
\examples{
data(exampleRanks)
data(examplePathways)
ranks <- sort(exampleRanks, decreasing=TRUE)
es <- calcGseaStatCumulative(ranks, na.omit(match(examplePathways[[1]], names(ranks))), 1)
}