% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea-package.R
\docType{data}
\name{exampleRanks}
\alias{exampleRanks}
\title{Example vector of gene-level statistics obtained for Th1 polarization.}
\description{
The data were obtained by doing differential expression between
Naive and Th1-activated states for GEO dataset GSE14308. The
exact script is available as system.file("gen_gene_ranks.R", package="fgsea")
}
