% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea.R
\name{fgsea}
\alias{fgsea}
\title{Runs preranked gene set enrichment analysis.}
\usage{
fgsea(pathways, stats, nperm, minSize = 1, maxSize = Inf, nproc = 1,
  gseaParam = 1)
}
\arguments{
\item{pathways}{List of gene sets to check.}

\item{stats}{Named vector of gene-level stats. Names should be the same as in 'pathways'}

\item{nperm}{Number of permutations to do. Minimial possible nominal p-value is about 1/nperm}

\item{minSize}{Minimal size of a gene set to test. All pathways below the threshold are excluded.}

\item{maxSize}{Maximal size of a gene set to test. All pathways above the threshold are excluded.}

\item{nproc}{Number of parallel processes to use.}

\item{gseaParam}{GSEA parameter value.}
}
\value{
A table with GSEA results.
}
\description{
Runs preranked gene set enrichment analysis.
}
\examples{
data(examplePathways)
data(exampleRanks)
\dontrun{
 fgseaRes <- fgsea(examplePathways, exampleRanks, nperm=10000, maxSize=500)
}
}

