% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgseaMultilevel.R
\name{multilevelError}
\alias{multilevelError}
\title{Calculates the expected error for the standard deviation of the P-value logarithm.}
\usage{
multilevelError(pval, sampleSize)
}
\arguments{
\item{pval}{P-value}

\item{sampleSize}{equivavlent to sampleSize in fgseaMultilevel}
}
\value{
The value of the expected error
}
\description{
Calculates the expected error for the standard deviation of the P-value logarithm.
}
\examples{
expectedError <- multilevelError(pval=1e-10, sampleSize=1001)
}
